feat: Migrate and standardize hotkey definitions
- Added default hotkey bindings in core defaults
- Implemented migration utility for legacy hotkey definitions
- Updated HotkeysManager to support migrating old hotkey configurations
- Simplified hotkey bindings customization in default extension
