+++ platform/core/src/classes/HotkeysManager.ts
@@ -1,6 +1,226 @@
 import objectHash from 'object-hash';
 import { hotkeys as mouseTrapAPI } from '../utils';
 import Hotkey from './Hotkey';
+import migrateOldHotkeyDefinitions from '../utils/hotkeys/migrateHotkeys';
+
+localStorage.setItem(
+  'hotkey-definitions',
+  JSON.stringify([
+    {
+      commandName: 'setToolActive',
+      commandOptions: {
+        toolName: 'Zoom',
+      },
+      keys: ['f'],
+      label: 'Zoom',
+      isEditable: true,
+    },
+    {
+      commandName: 'scaleUpViewport',
+      commandOptions: {},
+      keys: ['pp'],
+      label: 'Zoom In',
+      isEditable: true,
+    },
+    {
+      commandName: 'scaleDownViewport',
+      commandOptions: {},
+      keys: ['-'],
+      label: 'Zoom Out',
+      isEditable: true,
+    },
+    {
+      commandName: 'fitViewportToWindow',
+      commandOptions: {},
+      keys: ['='],
+      label: 'Zoom to Fit',
+      isEditable: true,
+    },
+    {
+      commandName: 'rotateViewportCW',
+      commandOptions: {},
+      keys: ['r'],
+      label: 'Rotate Right',
+      isEditable: true,
+    },
+    {
+      commandName: 'rotateViewportCCW',
+      commandOptions: {},
+      keys: ['l'],
+      label: 'Rotate Left',
+      isEditable: true,
+    },
+    {
+      commandName: 'flipViewportHorizontal',
+      commandOptions: {},
+      keys: ['h'],
+      label: 'Flip Horizontally',
+      isEditable: true,
+    },
+    {
+      commandName: 'flipViewportVertical',
+      commandOptions: {},
+      keys: ['v'],
+      label: 'Flip Vertically',
+      isEditable: true,
+    },
+    {
+      commandName: 'toggleCine',
+      commandOptions: {},
+      keys: ['c'],
+      label: 'Cine',
+    },
+    {
+      commandName: 'invertViewport',
+      commandOptions: {},
+      keys: ['i'],
+      label: 'Invert',
+      isEditable: true,
+    },
+    {
+      commandName: 'incrementActiveViewport',
+      commandOptions: {},
+      keys: ['right'],
+      label: 'Next Image Viewport',
+      isEditable: true,
+    },
+    {
+      commandName: 'decrementActiveViewport',
+      commandOptions: {},
+      keys: ['left'],
+      label: 'Previous Image Viewport',
+      isEditable: true,
+    },
+    {
+      commandName: 'updateViewportDisplaySet',
+      commandOptions: {
+        direction: -1,
+      },
+      keys: ['pageup'],
+      label: 'Previous Series',
+      isEditable: true,
+    },
+    {
+      commandName: 'updateViewportDisplaySet',
+      commandOptions: {
+        direction: 1,
+      },
+      keys: ['pagedown'],
+      label: 'Next Series',
+      isEditable: true,
+    },
+    {
+      commandName: 'nextStage',
+      commandOptions: {},
+      keys: ['.'],
+      label: 'Next Stage',
+      isEditable: true,
+    },
+    {
+      commandName: 'previousStage',
+      commandOptions: {},
+      keys: [','],
+      label: 'Previous Stage',
+      isEditable: true,
+    },
+    {
+      commandName: 'nextImage',
+      commandOptions: {},
+      keys: ['down'],
+      label: 'Next Image',
+      isEditable: true,
+    },
+    {
+      commandName: 'previousImage',
+      commandOptions: {},
+      keys: ['up'],
+      label: 'Previous Image',
+      isEditable: true,
+    },
+    {
+      commandName: 'firstImage',
+      commandOptions: {},
+      keys: ['home'],
+      label: 'First Image',
+      isEditable: true,
+    },
+    {
+      commandName: 'lastImage',
+      commandOptions: {},
+      keys: ['end'],
+      label: 'Last Image',
+      isEditable: true,
+    },
+    {
+      commandName: 'resetViewport',
+      commandOptions: {},
+      keys: ['space'],
+      label: 'Reset',
+      isEditable: true,
+    },
+    {
+      commandName: 'cancelMeasurement',
+      commandOptions: {},
+      keys: ['esc'],
+      label: 'Cancel Cornerstone Measurement',
+    },
+    {
+      commandName: 'setWindowLevel',
+      commandOptions: {
+        description: 'Soft tissue',
+        window: '400',
+        level: '40',
+      },
+      keys: ['1'],
+      label: 'W/L Preset 1',
+    },
+    {
+      commandName: 'setWindowLevel',
+      commandOptions: {
+        description: 'Lung',
+        window: '1500',
+        level: '-600',
+      },
+      keys: ['2'],
+      label: 'W/L Preset 2',
+    },
+    {
+      commandName: 'setWindowLevel',
+      commandOptions: {
+        description: 'Liver',
+        window: '150',
+        level: '90',
+      },
+      keys: ['3'],
+      label: 'W/L Preset 3',
+    },
+    {
+      commandName: 'setWindowLevel',
+      commandOptions: {
+        description: 'Bone',
+        window: '2500',
+        level: '480',
+      },
+      keys: ['4'],
+      label: 'W/L Preset 4',
+    },
+    {
+      commandName: 'setWindowLevel',
+      commandOptions: {
+        description: 'Brain',
+        window: '80',
+        level: '40',
+      },
+      keys: ['5'],
+      label: 'W/L Preset 5',
+    },
+    {
+      commandName: 'deleteActiveAnnotation',
+      commandOptions: {},
+      keys: ['backspace'],
+    },
+  ])
+);
 
 /**
  *
@@ -24,6 +244,11 @@ export class HotkeysManager {
   ) {
     this._servicesManager = servicesManager;
     this._commandsManager = commandsManager;
+
+    // Check for old hotkey definitions format and migrate if needed
+    migrateOldHotkeyDefinitions({
+      generateHash: this.generateHash,
+    });
   }
 
   /**
